Fixed CodeFactor issues manually
